reflection

+++ include/matrix.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   matrix.h                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jeada-si <jeada-si@student.42.fr>          +#+  +:+       +#+        */
+/*   By: lpaquatt <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/29 15:23:04 by jeada-si          #+#    #+#             */
-/*   Updated: 2024/06/10 12:38:56 by jeada-si         ###   ########.fr       */
+/*   Updated: 2024/06/11 18:02:26 by lpaquatt         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,8 +29,8 @@ t_mat	ft_mat_trans(t_mat A);
 t_mat	ft_mat_inv(t_mat A);
 int		ft_is_inv(t_mat A);
 t_tuple	ft_mat_prod_tup(t_mat A, t_tuple tuple);
-t_mat	ft_translation(int x, int y, int z);
-t_mat	ft_scaling(int x, int y, int z); 
+t_mat	ft_translation(double x, double y, double z);
+t_mat	ft_scaling(double x, double y, double z);
 t_mat	ft_mat_id(int size);
 
 #endif
